+++ Audibly.App/Helpers/Sentry.cs
@@ -1,0 +1,10 @@
+﻿// Author: rstewa · https://github.com/rstewa
+// Created: 06/26/2024
+// Updated: 06/26/2024
+
+namespace Audibly.App.Helpers;
+
+internal static class Sentry
+{
+    public const string Dsn = "sentry-dsn-url-goes-here";
+}